requirements.txt
requests
httpx[http2]
web3
python-dotenv
